+++ webrequestsimulator/internal/engine/types.go
@@ -1,0 +1,37 @@
+/*
+Copyright 2024.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+// Package engine implements the WebRequestCommitStatus reconcile simulator.
+// It is internal to webrequestsimulator (see Go internal directory rules).
+package engine
+
+import (
+	promoterv1alpha1 "github.com/argoproj-labs/gitops-promoter/api/v1alpha1"
+	"github.com/argoproj-labs/gitops-promoter/internal/webrequest"
+	"github.com/argoproj-labs/gitops-promoter/webrequestsimulator/types"
+)
+
+// Args carries everything the engine needs; the parent maps types.Input here.
+type Args struct {
+	WebRequestCommitStatus *promoterv1alpha1.WebRequestCommitStatus
+	PromotionStrategy      *promoterv1alpha1.PromotionStrategy
+	NamespaceMetadata      webrequest.NamespaceMetadata
+	HTTPResponse           *webrequest.HTTPResponse
+}
+
+type Result = types.Result
+
+type RenderedRequest = types.RenderedRequest
